+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - feat: hold update until the app creates a file signaling it is loaded - closes #177
 - feat: setup a wifi hotspot for when the a web-ui input is demanded and the device is not connected to any network - closes #169
 - feat: let the user upload directory content of the docker composition they are creating - closes #202
+- feat: use `pod_id` as the ssid of the wifi hotspot
 
 ## Version 1.1.0